+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.7.0
 ## Key tag
 - Added key tag for getting key from lists
+- Changed how tags are accessed and refactored the code
 # v0.6.0
 ## Added conditionals
 - Conditionals are now supported